security: keep httptest out of the agent binary on Go 1.27
