+++ security/pkg/credentialfetcher/plugin/mock.go
@@ -17,17 +17,8 @@
 package plugin
 
 import (
-	"fmt"
-	"net/http"
-	"net/http/httptest"
-	"net/url"
-	"os"
-	"sync"
-
 	"istio.io/istio/pkg/log"
 )
-
-const fakeTokenPrefix = "fake-token-"
 
 var mockcredLog = log.RegisterScope("mockcred", "Mock credential fetcher for istio agent")
 
@@ -56,68 +47,3 @@ func (p *MockPlugin) GetIdentityProvider() string {
 }
 
 func (p *MockPlugin) Stop() {}
-
-// MetadataServer mocks GCE metadata server.
-type MetadataServer struct {
-	server *httptest.Server
-
-	numGetTokenCall int
-	credential      string
-	mutex           sync.RWMutex
-}
-
-// StartMetadataServer starts a mock GCE metadata server.
-func StartMetadataServer() (*MetadataServer, error) {
-	ms := &MetadataServer{}
-	httpServer := httptest.NewServer(http.HandlerFunc(ms.getToken))
-	ms.server = httpServer
-	url, err := url.Parse(httpServer.URL)
-	if err != nil {
-		return nil, fmt.Errorf("parse URL failed: %v", err)
-	}
-	if err := os.Setenv("GCE_METADATA_HOST", url.Host); err != nil {
-		fmt.Printf("Error running os.Setenv: %v", err)
-		ms.Stop()
-		return nil, err
-	}
-	return ms, nil
-}
-
-func (ms *MetadataServer) setToken(t string) {
-	ms.mutex.Lock()
-	defer ms.mutex.Unlock()
-	ms.credential = t
-}
-
-// NumGetTokenCall returns the number of token fetching request.
-func (ms *MetadataServer) NumGetTokenCall() int {
-	ms.mutex.RLock()
-	defer ms.mutex.RUnlock()
-
-	return ms.numGetTokenCall
-}
-
-// Reset resets members to default values.
-func (ms *MetadataServer) Reset() {
-	ms.mutex.Lock()
-	defer ms.mutex.Unlock()
-
-	ms.numGetTokenCall = 0
-	ms.credential = ""
-}
-
-func (ms *MetadataServer) getToken(w http.ResponseWriter, req *http.Request) {
-	ms.mutex.Lock()
-	defer ms.mutex.Unlock()
-
-	ms.numGetTokenCall++
-	token := fmt.Sprintf("%s%d", fakeTokenPrefix, ms.numGetTokenCall)
-	if ms.credential != "" {
-		token = ms.credential
-	}
-	fmt.Fprint(w, token)
-}
-
-func (ms *MetadataServer) Stop() {
-	ms.server.Close()
-}
